+++ aldryn_newsblog/cms_wizards.py
@@ -7,8 +7,8 @@ from cms.wizards.forms import BaseFormMixin
 from cms.wizards.wizard_base import Wizard
 from cms.wizards.wizard_pool import wizard_pool
 
-from djangocms_text_ckeditor.html import clean_html
-from djangocms_text_ckeditor.widgets import TextEditorWidget
+from djangocms_text.html import clean_html
+from djangocms_text.widgets import TextEditorWidget
 from parler.forms import TranslatableModelForm
 
 from .cms_appconfig import NewsBlogConfig
@@ -76,6 +76,8 @@ class CreateNewsBlogArticleForm(BaseFormMixin, TranslatableModelForm):
         widgets = {'app_config': forms.Select()}
 
     def __init__(self, **kwargs):
+        for key in ("wizard_site", "wizard_request"):
+            kwargs.setdefault(key)
         super().__init__(**kwargs)
 
         # If there's only 1 (or zero) app_configs, don't bother show the