+++ aldryn_newsblog/tests/test_cms_wizards.py
@@ -3,6 +3,7 @@ from aldryn_newsblog.tests import NewsBlogTestCase
 
 
 class CreateNewsBlogArticleFormTestCase(NewsBlogTestCase):
+
     def get_form(self, has_content, has_permission):
         data = {'title': 'My super title', 'app_config': self.app_config.id}
         if has_content:
@@ -21,7 +22,7 @@ class CreateNewsBlogArticleFormTestCase(NewsBlogTestCase):
         self.assertEqual(article.content.get_plugins('en').count(), 1)
         plugin = article.content.get_plugins('en').get()
         self.assertEqual(plugin.plugin_type, 'TextPlugin')
-        self.assertEqual(plugin.djangocms_text_ckeditor_text.body, 'My super content')
+        self.assertEqual(plugin.djangocms_text_text.body, 'My super content')
 
     def test_article_is_saved_without_content_with_plugin_permission(self):
         form = self.get_form(has_content=False, has_permission=True)