+++ aldryn_newsblog/tests/test_feeds.py
@@ -1,16 +1,15 @@
 from datetime import timedelta
 
-from django.test import TransactionTestCase
 from django.urls import reverse
 from django.utils.timezone import now
 from django.utils.translation import override
 
 from aldryn_newsblog.feeds import CategoryFeed, LatestArticlesFeed, TagFeed
 
-from . import NewsBlogTestsMixin
+from . import NewsBlogTransactionTestCase
 
 
-class TestFeeds(NewsBlogTestsMixin, TransactionTestCase):
+class TestFeeds(NewsBlogTransactionTestCase):
 
     def test_latest_feeds(self):
         article = self.create_article()